+++ README.md
@@ -10,11 +10,6 @@ This differs by aiming to be less experimental than others, driven by industry u
 Many of the exposed types are a very thin layer above built in functionality.
 The goal is to provide all of the building blocks necessary to make concise, yet complex types.
 
-Additionally packaged with this lib is a JSON-schema validator that will act as a type-guard for appropriately typed object.
-Using this will mean run-time checking of user/network/external data, with compile time checking of logic flow.
-The primary purpose of packaging the JSON-schema validator is as an example of complexity of types this library is able to specify.
-In particular, this shows the library's solution to the conditional mapped types problem without overloading the global namespace.
-
 ```
 npm install --save-dev simplytyped
 ```
@@ -50,8 +45,6 @@ npm install --save-dev simplytyped
 **[Utils](#utils)**
 
 [Nullable](#nullable) - [NotNullable](#notnullable) - [NoInfer](#noinfer) - [Unknown](#unknown) - [Nominal](#nominal)
-
-**[Schema Validation](#schema-validation)**
 
 ## Conditionals
 
@@ -509,33 +502,3 @@ const obj = { a: '', b: 22 };
 const keys = objectKeys(obj); // Array<'a' | 'b'>
 ```
 
-## Schema Validation
-One of the easiest points of failure with the typescript type system is outside data.
-It is difficult to confirm that outside data matches the contract we have set within our typings.
-Using the common JSON-schema specification, we can check the validity of our runtime objects, while still having compile checking of logic validity.
-If we define our schemas as so:
-```ts
-const schema = {
-    type: 'object' as 'object',
-    properties: {
-        prop: {
-            type: 'string' as 'string'
-        }
-    }
-}
-```
-then pass in data:
-```ts
-// use `any` to pretend we don't know the type here.
-const runtimeData: any = {
-    prop: 'hello world'
-};
-```
-then `schemaIsValid(data, schema)` will give both compile time and run time checking of types.
-```ts
-if (schemaIsValid(data, schema)) {
-    type x = typeof data; // => { prop: string }
-} else {
-    throw new Error('Uh-oh, you gave me ill-formatted data!!');
-}
-```
