+++ src/impl.ts
@@ -1,0 +1,1 @@
+export * from './impl/objects';
